+++ src/index.ts
@@ -106,7 +106,6 @@ export * from './app/components/file-browser/terra-file-browser.service';
 export * from './app/components/file-browser/terra-file-browser.component';
 export * from './app/components/category-picker/terra-category-picker.component';
 export * from './app/components/category-picker/service/terra-category-picker-base.service';
-export * from './app/components/pager/data/terra-pager.data';
 export * from './app/models/terra-key-value.interface';
 export * from './app/components/category-picker/terra-category-picker.component';
 export * from './app/components/category-picker/data/category-data.interface';